feat: Add Restic password generator script
The Restic password generator script generates secure passwords suitable for use with Restic backups. It allows users to generate one or multiple passwords and optionally save them to a file. The script is customizable through command-line arguments, such as specifying the number of passwords to generate and the output file name. This feature will enhance the functionality of the Restic backup system.